test: sanitize Windows CI vitest runner output (#567) (thanks @erikpr1994)
