+++ test/vitest-global-setup.ts
@@ -1,0 +1,5 @@
+import { installWindowsCIOutputSanitizer } from "./windows-ci-output-sanitizer";
+
+export default function globalSetup() {
+  installWindowsCIOutputSanitizer();
+}
